+++ spec/spec/container.js
@@ -147,20 +147,20 @@ describe('Container', function() {
   describe('image()', function() {
     it('should increase children by 1', function() {
       var initial = draw.children().length
-      draw.image(imageUrl, 100, 100)
+      draw.image(imageUrl)
       expect(draw.children().length).toBe(initial + 1)
     })
     it('should create a rect', function() {
-      expect(draw.image(imageUrl, 100, 100).type).toBe('image')
+      expect(draw.image(imageUrl).type).toBe('image')
     })
     it('should create an instance of SVG.Rect', function() {
-      expect(draw.image(imageUrl, 100, 100) instanceof SVG.Image).toBe(true)
+      expect(draw.image(imageUrl) instanceof SVG.Image).toBe(true)
     })
     it('should be an instance of SVG.Shape', function() {
-      expect(draw.image(imageUrl, 100, 100) instanceof SVG.Shape).toBe(true)
+      expect(draw.image(imageUrl) instanceof SVG.Shape).toBe(true)
     })
     it('should be an instance of SVG.Element', function() {
-      expect(draw.image(imageUrl, 100, 100) instanceof SVG.Element).toBe(true)
+      expect(draw.image(imageUrl) instanceof SVG.Element).toBe(true)
     })
   })
 
@@ -173,11 +173,11 @@ describe('Container', function() {
     it('creates a text element', function() {
       expect(draw.text(loremIpsum).type).toBe('text')
     })
-    it('creates an instance of SVG.Rect', function() {
+    it('creates an instance of SVG.Text', function() {
       expect(draw.text(loremIpsum) instanceof SVG.Text).toBe(true)
     })
-    it('is an instance of SVG.Shape', function() {
-      expect(draw.text(loremIpsum) instanceof SVG.Shape).toBe(true)
+    it('is an instance of SVG.Parent', function() {
+      expect(draw.text(loremIpsum) instanceof SVG.Parent).toBe(true)
     })
     it('is an instance of SVG.Element', function() {
       expect(draw.text(loremIpsum) instanceof SVG.Element).toBe(true)
@@ -196,8 +196,8 @@ describe('Container', function() {
     it('creates an instance of SVG.Rect', function() {
       expect(draw.plain(loremIpsum) instanceof SVG.Text).toBe(true)
     })
-    it('is an instance of SVG.Shape', function() {
-      expect(draw.plain(loremIpsum) instanceof SVG.Shape).toBe(true)
+    it('is an instance of SVG.Parent', function() {
+      expect(draw.plain(loremIpsum) instanceof SVG.Parent).toBe(true)
     })
     it('is an instance of SVG.Element', function() {
       expect(draw.plain(loremIpsum) instanceof SVG.Element).toBe(true)