+++ spec/spec/selector.js
@@ -4,11 +4,11 @@ describe('Selector', function() {
     it('gets an element\'s instance by id', function() {
       var rect = draw.rect(111, 333)
       
-      expect(SVG.get(rect.attr('id'))).toBe(rect)
+      expect(SVG.get(rect.id())).toBe(rect)
     })
     it('makes all the element\'s methods available', function() {
       var element = draw.group()
-        , got = SVG.get(element.attr('id'))
+        , got = SVG.get(element.id())
       
       expect(got.attr()).toEqual(element.attr())
     })
@@ -36,8 +36,8 @@ describe('Selector', function() {
     it('gets all elements with a given class name', function() {
       expect(SVG.select('rect.selectable-element').valueOf()).toEqual([e1, e3, e5])
     })
-    it('returns an instance of SVG.Set', function() {
-      expect(SVG.select('rect.selectable-element') instanceof SVG.Set).toBe(true)
+    it('returns an Array', function() {
+      expect(SVG.select('rect.selectable-element') instanceof Array).toBe(true)
     })
   })