+++ src/attr.js
@@ -28,16 +28,11 @@ SVG.extend(SVG.Element, {
         parseFloat(v) : v
     
     } else {
-      // BUG FIX: some browsers will render a stroke if a color is given even though stroke width is 0
-      if (a == 'stroke-width')
-        this.attr('stroke', parseFloat(v) > 0 ? this._stroke : null)
-      else if (a == 'stroke')
-        this._stroke = v
 
       // convert image fill and stroke to patterns
       if (a == 'fill' || a == 'stroke') {
         if (SVG.regex.isImage.test(v))
-          v = this.doc().defs().image(v, 0, 0)
+          v = this.doc().defs().image(v)
 
         if (v instanceof SVG.Image)
           v = this.doc().defs().pattern(0, 0, function() {